split error handling code to two file
discard old validator code
