+++ portal/src/util/error.ts
@@ -1,28 +1,7 @@
 import { ApolloError } from "@apollo/client";
 import { GraphQLError } from "graphql";
 import { nonNullable } from "./types";
-import { Values } from "@oursky/react-messageformat";
-
-// union type of different kind of violation
-export type Violation = RequiredViolation | GeneralViolation;
-
-interface RequiredViolation {
-  kind: "required";
-  location: string;
-  missingField: string[];
-}
-
-interface GeneralViolation {
-  kind: "general";
-  location: string;
-}
-
-// list of violation kind recognized
-const violationKinds = ["required", "general"];
-type ViolationKind = Violation["kind"];
-function isViolationKind(value?: string): value is ViolationKind {
-  return value != null && violationKinds.includes(value);
-}
+import { Violation } from "./validation";
 
 // expected data shape of error extension from backend
 interface RequiredErrorCauseDetails {
@@ -63,20 +42,11 @@ interface APIValidationError {
 // union type of api errors, depend on reason
 type APIError = APIValidationError;
 
-type ViolationSelector = (violation: Violation) => boolean;
-type ViolationSelectors<Key extends string> = Record<Key, ViolationSelector>;
-
 function isAPIError(value?: { [key: string]: any }): value is APIError {
   if (value == null) {
     return false;
   }
   return "errorName" in value && "info" in value && "reason" in value;
-}
-
-function defaultFormatErrorMessageList(
-  errorMessages: string[]
-): string | undefined {
-  return errorMessages.length === 0 ? undefined : errorMessages.join("\n");
 }
 
 function extractViolationFromErrorCause(cause: ErrorCause): Violation | null {
@@ -125,78 +95,4 @@ export function parseError(error: unknown): Violation[] {
 
   // unrecognized error
   return [];
-}
-
-// default seclectors
-export function makeMissingFieldSelector(
-  locationPrefix: string,
-  fieldName: string
-): ViolationSelector {
-  return function (violation: Violation) {
-    if (violation.kind !== "required") {
-      return false;
-    }
-    if (!violation.location.startsWith(locationPrefix)) {
-      return false;
-    }
-    return violation.missingField.includes(fieldName);
-  };
-}
-
-export function combineSelector(
-  selectorList: ViolationSelector[]
-): ViolationSelector {
-  return function (violation: Violation) {
-    for (const selector of selectorList) {
-      if (selector(violation)) {
-        return true;
-      }
-    }
-    return false;
-  };
-}
-
-export function violationSelector<Key extends string>(
-  violations: Violation[],
-  violationSelectors: ViolationSelectors<Key>
-): Record<Key, Violation[]> {
-  const violationMap = Object.entries(violationSelectors).reduce<
-    Partial<Record<Key, Violation[]>>
-  >((violationMap, [key, selector]) => {
-    violationMap[key as Key] = violations.filter(selector as ViolationSelector);
-    return violationMap;
-  }, {});
-  return violationMap as Record<Key, Violation[]>;
-}
-
-function violationFormatter(
-  fieldNameId: string,
-  violation: Violation,
-  renderToString: (messageId: string, values?: Values) => string
-): string | undefined {
-  switch (violation.kind) {
-    case "required":
-      return renderToString("required-field-missing", {
-        fieldName: renderToString(fieldNameId),
-      });
-    default:
-      return undefined;
-  }
-}
-
-export function errorFormatter(
-  fieldNameId: string,
-  violations: Violation[],
-  renderToString: (messageId: string, values?: Values) => string,
-  formatErrorMessageList: (
-    errorMessages: string[]
-  ) => string | undefined = defaultFormatErrorMessageList
-): string | undefined {
-  return formatErrorMessageList(
-    violations
-      .map((violation) =>
-        violationFormatter(fieldNameId, violation, renderToString)
-      )
-      .filter(nonNullable)
-  );
 }
