+++ portal/src/util/validation.ts
@@ -1,58 +1,106 @@
-export interface ValidationRule<Data, ErrorState> {
-  inputKey: keyof Data;
-  errorKey: keyof ErrorState;
-  condition: (input: any) => boolean;
-  errorMessageId: string;
+import { Values } from "@oursky/react-messageformat";
+import { nonNullable } from "./types";
+
+// union type of different kind of violation
+export type Violation = RequiredViolation | GeneralViolation;
+
+interface RequiredViolation {
+  kind: "required";
+  location: string;
+  missingField: string[];
 }
 
-export type ErrorResult<ErrorState> = Partial<
-  { [key in keyof ErrorState]: string }
->;
+interface GeneralViolation {
+  kind: "general";
+  location: string;
+}
 
-export function validateInput<Data, ErrorState>(
-  data: Data,
-  validationRules: ValidationRule<Data, ErrorState>[]
-): {
-  errorResult: ErrorResult<ErrorState>;
-  isValid: boolean;
-} {
-  let isValid = true;
-  const errorResult: ErrorResult<ErrorState> = {};
-  validationRules.forEach(
-    (validationRule: ValidationRule<Data, ErrorState>) => {
-      if (!validationRule.condition(data[validationRule.inputKey])) {
-        errorResult[validationRule.errorKey] = validationRule.errorMessageId;
-        isValid = false;
+// list of violation kind recognized
+const violationKinds = ["required", "general"];
+type ViolationKind = Violation["kind"];
+export function isViolationKind(value?: string): value is ViolationKind {
+  return value != null && violationKinds.includes(value);
+}
+
+type ViolationSelector = (violation: Violation) => boolean;
+type ViolationSelectors<Key extends string> = Record<Key, ViolationSelector>;
+
+function defaultFormatErrorMessageList(
+  errorMessages: string[]
+): string | undefined {
+  return errorMessages.length === 0 ? undefined : errorMessages.join("\n");
+}
+
+// default seclectors
+export function makeMissingFieldSelector(
+  locationPrefix: string,
+  fieldName: string
+): ViolationSelector {
+  return function (violation: Violation) {
+    if (violation.kind !== "required") {
+      return false;
+    }
+    if (!violation.location.startsWith(locationPrefix)) {
+      return false;
+    }
+    return violation.missingField.includes(fieldName);
+  };
+}
+
+export function combineSelector(
+  selectorList: ViolationSelector[]
+): ViolationSelector {
+  return function (violation: Violation) {
+    for (const selector of selectorList) {
+      if (selector(violation)) {
+        return true;
       }
     }
-  );
-  return { errorResult, isValid };
-}
-
-export function isValidEmail(email: string): boolean {
-  const emailRegExp = /^.+@.+$/;
-  return emailRegExp.test(email);
-}
-
-export function isValidEmailDomain(domain: string): boolean {
-  return domain !== "";
-}
-
-export function isErrorResult<ErrorState>(
-  result: any
-): result is ErrorResult<ErrorState> {
-  if (!(result instanceof Object)) {
     return false;
-  }
-  return Object.keys(result).length > 0;
+  };
 }
 
-export function renderErrorMessage(
-  renderToString: (messageId: string) => string,
-  errorMessageId?: string
+export function violationSelector<Key extends string>(
+  violations: Violation[],
+  violationSelectors: ViolationSelectors<Key>
+): Record<Key, Violation[]> {
+  const violationMap = Object.entries(violationSelectors).reduce<
+    Partial<Record<Key, Violation[]>>
+  >((violationMap, [key, selector]) => {
+    violationMap[key as Key] = violations.filter(selector as ViolationSelector);
+    return violationMap;
+  }, {});
+  return violationMap as Record<Key, Violation[]>;
+}
+
+function violationFormatter(
+  fieldNameId: string,
+  violation: Violation,
+  renderToString: (messageId: string, values?: Values) => string
 ): string | undefined {
-  if (errorMessageId == null) {
-    return undefined;
+  switch (violation.kind) {
+    case "required":
+      return renderToString("required-field-missing", {
+        fieldName: renderToString(fieldNameId),
+      });
+    default:
+      return undefined;
   }
-  return renderToString(errorMessageId);
+}
+
+export function errorFormatter(
+  fieldNameId: string,
+  violations: Violation[],
+  renderToString: (messageId: string, values?: Values) => string,
+  formatErrorMessageList: (
+    errorMessages: string[]
+  ) => string | undefined = defaultFormatErrorMessageList
+): string | undefined {
+  return formatErrorMessageList(
+    violations
+      .map((violation) =>
+        violationFormatter(fieldNameId, violation, renderToString)
+      )
+      .filter(nonNullable)
+  );
 }
